Added support for: bit.ly, goo.gl, ndurl.com, kl.am, qr.cx. Removed tr.im. Added JSONKit support. Fixed some bugs in OSX 10.6. There is still an issue with the test app in 64 bit.

+++ Classes/Core/USIncludes.h
@@ -5,11 +5,10 @@
 + (id)allocWithZone:(NSZone *)zone{@synchronized(self){if( y == nil) { y = [super allocWithZone:zone]; return y ;}} return nil;} \
 - (id)copyWithZone:(NSZone *)zone{return self;} \
 - (id)retain{return self;} \
-- (unsigned)retainCount{return UINT_MAX;} \
+- (NSUInteger)retainCount{return UINT_MAX;} \
 - (void)release{} \
 - (id)autorelease{return self;}
 
 #define kUSShrinkChoiceDefaultsKey @"Shrinker"
-
 #define kSDShortenURLGLobalShortcutKey @"kSDShortenURLGLobalShortcutKey"
 #define GetRandom(_min,_max) ((rand() % ((_max) - (_min) - 1)) + (_min))
